requirements.txt
# Discord bot dependencies
discord.py>=2.3.0
aiohttp>=3.8.0

# Database drivers
pymongo>=4.3.0
aiomysql>=0.2.0
aiosqlite>=0.19.0

# Configuration and utilities
python-dotenv>=1.0.0
pydantic>=2.0.0
PyYAML>=6.0

# Testing dependencies
pytest>=7.0.0
pytest-asyncio>=0.21.0
pytest-mock>=3.10.0

# Development dependencies
black>=23.0.0
flake8>=6.0.0
mypy>=1.0.0